renamed `test` command
